requirements.txt
Django>=3.2.12,<4.3
djangorestframework>=3.13.1,<3.14
djangorestframework-simplejwt>=5.1.0,<5.2.0
Pillow>=9.0.1,<9.1
drf-spectacular>=0.22.0,<0.25.0
drf-spectacular-sidecar>=2022.4.1,<2022.5.1
django-mptt>=0.13.4,<0.15.0
django-filter>=21.1,<22.1
drf-extra-fields>=3.4.0,<3.5.0
gunicorn>=20.1.0,<21.1.0
flake8>=4.0.1,<4.1
psycopg2>=2.9.3,<2.10
uWSGI>=2.0.20,<2.1
